+++ src/main/java/spinnery/widget/WDynamicImage.java
@@ -8,7 +8,7 @@ public class WDynamicImage extends WWidget implements WClient {
 
 	protected int position = 0;
 
-	public WDynamicImage(WAnchor anchor, double positionX, double positionY, double positionZ, double sizeX, double sizeY, WInterface linkedPanel, Identifier... textures) {
+	public WDynamicImage(WAnchor anchor, int positionX, int positionY, int positionZ, int sizeX, int sizeY, WInterface linkedPanel, Identifier... textures) {
 		setInterface(linkedPanel);
 
 		setAnchor(anchor);
